Continuo otra vez con SDL2

+++ test/SDL2_BOX2D/spriteMotion/src/Constants.hpp
@@ -8,11 +8,11 @@ namespace consts {
     static constexpr int CAMERA_HEIGHT_PX = 600;
     static constexpr int CAMERA_MIDDLE_HEIGHT_PX = CAMERA_HEIGHT_PX / 2;
     
-    static constexpr float32 WORLD_WIDTH_PX = 1679;
-    static constexpr float32 WORLD_MIDDLE_WIDTH_PX = WORLD_WIDTH_PX / 2;
+    static constexpr int WORLD_WIDTH_PX = 1679;
+    static constexpr int WORLD_MIDDLE_WIDTH_PX = WORLD_WIDTH_PX / 2;
     
-    static constexpr float32 WORLD_HEIGHT_PX = 600;
-    static constexpr float32 WORLD_MIDDLE_HEIGHT_PX = WORLD_HEIGHT_PX / 2;
+    static constexpr int WORLD_HEIGHT_PX = 600;
+    static constexpr int WORLD_MIDDLE_HEIGHT_PX = WORLD_HEIGHT_PX / 2;
     
     static constexpr int DELAY_BETWEEN_ANIMATIONS = 10;
     static constexpr int CAMERA_MARGIN_PX = 100;
